Use the .xcassets file for filetype images.
Doing this (instead of having the images floating in the app bundle)
allows the 6 plus (at least in the simulator) to load the 2x images,
instead of defaulting to the 1x. I’m doing this because I’m too lazy to
look for a new set of filetype icons big enough for 3x.